Generate ROOT dictionary for TruthDerived column
