+++ src/RarexsecDictLinkDef.h
@@ -1,0 +1,8 @@
+#ifdef __CLING__
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ namespace proc;
+#pragma link C++ struct proc::TruthDerived+;
+#endif
